working

+++ include/RungeKutta.cuh
@@ -1,3 +1,3 @@
-__host__ void RungeKutta (float *q0, float timestep, float *PlanetMasses, float *q1, int nb, int *feelothers);
-__host__ void TranslatePlanetRK5 (float *qold, float c1, float c2, float c3, float c4, float c5, float *qnew, int nb);
-__host__ void DerivMotionRK5 (float *q_init, float *PlanetMasses, float *deriv, int nb, float dt, int *feelothers);
+__host__ void RungeKutta (double *q0, float timestep, double *PlanetMasses, double *q1, int nb, int *feelothers);
+__host__ void TranslatePlanetRK5 (double *qold, double c1, double c2, double c3, double c4, double c5, double *qnew, int nb);
+__host__ void DerivMotionRK5 (double *q_init, double *PlanetMasses, double *deriv, int nb, float dt, int *feelothers);
